requirements.txt
pillow
pytesseract
python-docx
reportlab